+++ backend/swagger/docs/docs.go
@@ -2789,11 +2789,15 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "disableOfficePreviewExt": {
-                    "description": "comma separated list of file extensions to disable office preview for",
+                    "description": "deprecated: use disablePreviewExt instead",
                     "type": "string"
                 },
                 "disableOnlyOfficeExt": {
-                    "description": "comma separated list of file extensions to disable onlyoffice preview for",
+                    "description": "list of file extensions to disable onlyoffice editor for",
+                    "type": "string"
+                },
+                "disablePreviewExt": {
+                    "description": "comma separated list of file extensions to disable preview for",
                     "type": "string"
                 },
                 "disableQuickToggles": {
@@ -2811,6 +2815,10 @@ const docTemplate = `{
                 "disableUpdateNotifications": {
                     "description": "disable update notifications banner for admin users",
                     "type": "boolean"
+                },
+                "disableViewingExt": {
+                    "description": "comma separated list of file extensions to disable viewing for",
+                    "type": "string"
                 },
                 "editorQuickSave": {
                     "description": "show quick save button in editor",
@@ -3051,11 +3059,15 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "disableOfficePreviewExt": {
-                    "description": "comma separated list of file extensions to disable office preview for",
+                    "description": "deprecated",
                     "type": "string"
                 },
                 "disableOnlyOfficeExt": {
-                    "description": "comma separated list of file extensions to disable onlyoffice preview for",
+                    "description": "deprecated",
+                    "type": "string"
+                },
+                "disablePreviewExt": {
+                    "description": "comma separated list of file extensions to disable preview for",
                     "type": "string"
                 },
                 "disableQuickToggles": {
@@ -3072,6 +3084,10 @@ const docTemplate = `{
                 "disableUpdateNotifications": {
                     "description": "disable update notifications",
                     "type": "boolean"
+                },
+                "disableViewingExt": {
+                    "description": "comma separated list of file extensions to disable viewing for",
+                    "type": "string"
                 },
                 "editorQuickSave": {
                     "description": "show quick save button in editor",