+++ frontend/src/store/mutations.js
@@ -320,7 +320,6 @@ export const mutations = {
           "dateFormat",
           "themeColor",
           "quickDownload",
-          "disableOnlyOfficeExt",
           "preview",
           "stickySidebar",
           "singleClick",